+++ src/plotery.js
@@ -1,13 +1,30 @@
 export { Chart } from './chart.component';
 export { Surface } from './surface.component';
+
 export { CartesianAxis } from './cartesian/cartesian-axis.component';
 export { LinearAxis } from './cartesian/linear-axis.component';
 export { LogAxis } from './cartesian/log-axis.component';
+
 export { CardinalLine } from './cartesian/cardinal-line.component';
 export { LinearLine } from './cartesian/linear-line.component';
+export { CartesianLine } from './cartesian/cartesian-line.component';
 export { BarLine } from './cartesian/bar-line.component';
+
+export { PolarAxis } from './polar/polar-axis.component';
 export { RadialAxis } from './polar/radial-axis.component';
 export { AngularAxis } from './polar/angular-axis.component';
+
 export { PolarLine } from './polar/polar-line.component';
 export { PolarSector } from './polar/polar-sector.component';
+
 export { Zoom } from './interactions/zoom.component';
+
+export { linear } from './math/interpolations/linear';
+export { cardinal } from './math/interpolations/cardinal';
+export { linearScaler } from './math/scalers/linear-scaler';
+export { logScaler } from './math/scalers/log-scaler';
+export { estimateUniformStep } from './math/axes/estimate-uniform-step';
+export { estimateLinearStep } from './math/axes/estimate-linear-step';
+export { estimateLogStep } from './math/axes/estimate-log-step';
+export { generateLinearTicks } from './math/axes/generate-linear-ticks';
+export { generateLogTicks } from './math/axes/generate-log-ticks';